Changes to displays as well as power button addition
