+++ lib/AuroraExtension.py
@@ -7,6 +7,7 @@ import os, sys
 import logging
 import time
 import numpy as np
+import datetime
 
 
 class AuroraExtension:
@@ -186,6 +187,7 @@ class AuroraExtension:
                 r = 800 / float(pixelImageWidth)
                 dim = (800, int(pixelImageHeight * r))
                 pixelImage = cv2.resize(pixelImage, dim, interpolation=cv2.INTER_AREA)
+                
 
             cv2.imwrite(filepath, pixelImage)
             # self.log("Saved PixelImage")
@@ -226,64 +228,74 @@ class AuroraExtension:
 
         return [False, False]
 
-    def autocrop(self, image, threshold=0):
-        """Crops any edges below or equal to threshold
-        Crops blank image to 1x1.
-        Returns cropped image.
-        """
-        if len(image.shape) == 3:
-            flatImage = np.max(image, 2)
-        else:
-            flatImage = image
-        assert len(flatImage.shape) == 2
-
-        rows = np.where(np.max(flatImage, 0) > threshold)[0]
-        if rows.size:
-            cols = np.where(np.max(flatImage, 1) > threshold)[0]
-            image = image[cols[0] : cols[-1] + 1, rows[0] : rows[-1] + 1]
-        else:
-            image = image[:1, :1]
-
-        return image
-
-    def takeScreenShot(self, filepath, autocrop=False):
-        # logging.info("Taking Screenshot in Aurora Extension")
-        ret, self.current_frame = self.getFrame()
+    
+    
+    # Take a screenshot and put the outline over it, use the border edges if neccessary
+    def takeScreenShot(self, filepath, autocrop=False,aspectCrop=False, borderEdges = [0,0,0,0]):
         screenshot_frame = self.current_frame
 
         if autocrop != False:
             screenshot_frame = self.autocrop(self.current_frame, autocrop)
             # logging.info(f"Cropped Screenshot to {screenshot_frame.shape}")
 
+        if(aspectCrop != False):
+            vid_h, vid_w, channels = screenshot_frame.shape
+            aspectRatioGap = vid_h - (vid_w / self.aspectRatio)
+            aspectRatioGap = aspectRatioGap / 2 # since we want to put some top and bottom
+            aspectRatioGap = int(aspectRatioGap)
+            #aspectRatioGap = 50
+            borderEdges = [0,aspectRatioGap,0,aspectRatioGap]
+            #screenshot_frame = self.aspectCrop(self.current_frame,aspectCrop)
+            self.vid_h, self.vid_w, self.channels = screenshot_frame.shape
+
+
+
         self.vid_h, self.vid_w, self.channels = screenshot_frame.shape
-        widthPixels = int(self.vid_w * (self.percent / 100)) + 1
-        heightPixels = int(self.vid_h * (self.percent / 100) * 2) + 1
 
-        colour = (0, 0, 255)
+        #This is for the outline starting and ending points
+        borderGap = 2 #distance away from the border so we can still see the lines
+        self.log(borderEdges)
+        borderTop = 0 + borderGap - 2 + borderEdges[1]
+        borderLeft = 0 + borderGap - 2 + borderEdges[0]
+        borderBottom = self.vid_h-borderGap - borderEdges[3]
+        borderRight = self.vid_w - borderGap - borderEdges[2]
+        #133,0,345,638
+        
+        #border sizes
+        widthPixels = int(borderRight * (self.percent / 100)) + 1
+        heightPixels = int(borderBottom * (self.percent / 100) * 2) + 1
 
+       
+        if(widthPixels < 15):
+            widthPixels = 15
+        if(heightPixels < 12):
+            heightPixels = 12
+
+        borderColour = (0, 0,255)
+        
         # top
         screenshot_frame = cv2.rectangle(
-            screenshot_frame, (0, 0), (self.vid_w, heightPixels), (0, 0, 255), 1
+            screenshot_frame, (borderLeft,borderTop), (borderRight, borderTop+heightPixels), borderColour, 1
         )
 
         # bottom
         screenshot_frame = cv2.rectangle(
             screenshot_frame,
-            (0, self.vid_h - heightPixels),
-            (self.vid_w, self.vid_h),
-            (0, 0, 255),
+            (0, borderBottom - heightPixels),
+            (borderRight, borderBottom),
+            borderColour,
             1,
         )
         # left
         screenshot_frame = cv2.rectangle(
-            screenshot_frame, (0, 0), (widthPixels, self.vid_h), (0, 0, 255), 1
+            screenshot_frame, (borderLeft,borderTop), (widthPixels, borderBottom), borderColour, 1
         )
         # right
         screenshot_frame = cv2.rectangle(
             screenshot_frame,
-            (self.vid_w - widthPixels, 0),
-            (self.vid_w, self.vid_h),
-            (0, 0, 255),
+            (borderRight - widthPixels, borderTop),
+            (borderRight, borderBottom),
+            borderColour,
             1,
         )
 
@@ -298,13 +310,91 @@ class AuroraExtension:
             logging.debug("DEBUG {} : {}".format(self.Name, log_string))
 
     # This class runs the visualisation (mandatory)
-    def visualise(self):
-
+    def visualiseFrame(self,frame):
+        self.current_frame = frame
         try:
             # In your inheritted class run things here
-            self.example = True
+            self.vid_h, self.vid_w, self.channels = self.current_frame.shape
+
+            if self.vid_h <= 1 and self.vid_w <= 1:
+                if len([pixel for pixel in self.pixels if pixel != (0, 0, 0)]) > 0:
+                    self.pixels.fill((0, 0, 0))
+                    self.pixels.show()
+                return
+
+            widthPixels = int(self.vid_w * (self.percent / 100)) + 1
+            heightPixels = int(self.vid_h * (self.percent / 100) * 2) + 1
+
+            sectionTop = self.current_frame[0:heightPixels, 0 : self.vid_w]
+            sectionBottom = self.current_frame[
+                self.vid_h - heightPixels : self.vid_h, 0 : self.vid_w
+            ]
+
+            sectionLeft = self.current_frame[0 : self.vid_h, 0:widthPixels]
+            sectionRight = self.current_frame[
+                0 : self.vid_h, self.vid_w - widthPixels : self.vid_w
+            ]
+
+            # get shape
+            h, w, c = sectionTop.shape
+            hs = 1
+            ws = self.pixelsTop
+            topSize = 1 if self.pixelsTop == 0 else self.pixelsTop
+            resizedTop = cv2.resize(sectionTop, (topSize, hs), interpolation=cv2.INTER_AREA)
+            bottomSize = 1 if self.pixelsBottom == 0 else self.pixelsBottom
+            resizedBottom = cv2.resize(
+                sectionBottom, (bottomSize, hs), interpolation=cv2.INTER_AREA
+            )
+
+            # get shape for sides
+            h, w, c = sectionLeft.shape
+            hs = self.pixelsLeft
+            ws = 1
+            leftSize = 1 if self.pixelsLeft == 0 else self.pixelsLeft
+            resizedLeft = cv2.resize(sectionLeft, (ws, leftSize), interpolation=cv2.INTER_AREA)
+            rightSize = 1 if self.pixelsRight == 0 else self.pixelsRight
+            resizedRight = cv2.resize(sectionRight, (ws, rightSize), interpolation=cv2.INTER_AREA)
+
+            # self.log(f"ResizeTime: {datetime.datetime.now()-stopwatchStartTime}")
+            # stopwatchStartTime = datetime.datetime.now()
+            # Populate LEDs
+            startPoint = 0
+            for i in range(self.pixelsLeft):            
+                B, G, R = resizedLeft[i][0]
+                self.pixels[self.pixelsLeft - (startPoint + i) - 1] = (R, G, B)
+            
+            startShowTime = datetime.datetime.now()
+            startPoint += self.pixelsLeft
+            for i in range(self.pixelsTop):
+                B, G, R = resizedTop[0][i]
+                self.pixels[startPoint + i] = (R, G, B)
+
+            startPoint += self.pixelsTop
+            for i in range(self.pixelsRight):
+                B, G, R = resizedRight[i][0]
+                self.pixels[startPoint + i] = (R, G, B)
+
+            startPoint += self.pixelsRight
+            for i in range(self.pixelsBottom):
+                B, G, R = (0, 0, 0)
+                if(any(val > self.darkThreshhold for val in resizedBottom[0][i])):
+                    B, G, R = resizedBottom[0][i]
+                self.pixels[startPoint + self.pixelsBottom - i - 1] = (R, G, B)
+
+            for key,test_cols in enumerate(self.pixels):
+                if(all(val < self.darkThreshhold for val in test_cols)):
+                    self.pixels[key] = (0,0,0)
+            
+            # self.log(f"DisplayTime: {datetime.datetime.now()-stopwatchStartTime}")
+            # self.log(f"Total time taken: {datetime.datetime.now()-totalStartTime}")
+            self.pixels.show()
+
+            # self.log(f"Total time taken: {datetime.datetime.now()-totalStartTime}")
+            # self.log("----------------------")
+            # visualise!
+            self.count += 1
         except Exception as e:
-            print("Error: {}".format(e))
+            self.log(f"Error: {str(e)}")
 
     # Returns the current FPS of the application ( not mandatory)
     def showFPS(self):
