+++ webserver/static/js/aurora-generic.js
@@ -1,5 +1,10 @@
 currentAjaxRequest = null; // Stores our current ajax request so we can cancel it if we do another
 
+function togglePower()
+{
+    make_AJAX_Call("/toggleEnable",{$('#powerToggle').toggleClass("color-theme")})
+}
+
 function make_AJAX_Call(url, data_dict, callback_function = false) {
 
     return_result = false;
