+++ webserver/templates/header.html
@@ -48,8 +48,10 @@
 
         <div class="page-title page-title-fixed">
             <h1 class="font-24">Aurora</h1>
+            <a href="#" id="toggle-power-header" onclick="togglePower()" class="page-title-icon shadow-xl bg-theme " ><i
+                    class="fa {%if enabled == True %}fa-power-off{% else %}fa-power-on{% endif %}" id="powerToggle" ></i></a>
             <a href="/configure/" class="page-title-icon shadow-xl bg-theme color-theme" data-menu=""><i
-                    class="fa fa-wrench"></i></a>
+                class="fa fa-wrench"></i></a>
             <a href="#" class="page-title-icon shadow-xl bg-theme color-theme show-on-theme-light" data-toggle-theme><i
                     class="fa fa-moon"></i></a>
             <a href="#" class="page-title-icon shadow-xl bg-theme color-theme show-on-theme-dark" data-toggle-theme><i
